work in progress

+++ src/components/courts/Courts.tsx
@@ -1,4 +1,5 @@
 import './courts.css';
+import { useState } from "react";
 import { Rows } from './Rows';
 import { Header } from './Header';
 import { ReservationItem, NormalizedReservationItem, User } from '../../types';
@@ -16,6 +17,7 @@ type Props = {
 const hours = [8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22];
 
 export function Courts(props: Props) {
+    const [disabled, setDisabled] = useState(false);
     const user_id = useSelector((state: RootState) => state.auth._id);
     const { club_id } = props;
     const date = new Date().toISOString().split('T')[0];
@@ -30,42 +32,43 @@ export function Courts(props: Props) {
     };
 
     filteredReservations.map(item => item.user_name = getUserName(item.user_id));
-    console.log({filteredReservations})
 
     function deleteReservation(slot: HTMLElement) {
-        console.log('click on delete', slot);
-
-        slot.classList.add('loading');
+        if (disabled) {
+            return;
+        }
 
         const reservationId = slot.dataset.reservation_id;
-
-        console.log(reservationId)
 
         fetch(`/api/reservations?reservation_id=${reservationId}&club_id=${club_id}`, {
             method: 'DELETE'
         })
         .then((response) => response.json())
         .then(json => {
-            slot.classList.remove('loading');
-            console.log(json)
             if (json.error) {
                 console.error(json.error)
             } else {
-                console.log('Reservation deleted');
                 if (json.data) {
-                    //console.log(json.data);
                     props.setReservations(json.data);
                 }
             }
+        })
+        .finally(() => {
+            setDisabled(false);
         });
 
     }
 
     function clickHandler (event: React.MouseEvent) {
+        if (disabled) {
+            return;
+        }
+
+        setDisabled(true);
         if (event.target instanceof HTMLElement) {
             const slot = event.target;
+
             if (slot.classList.contains('reserved')) {
-                console.log('no click allowed');
                 return;
             }
 
@@ -98,15 +101,16 @@ export function Courts(props: Props) {
             .then(json => {
                 slot.classList.remove('loading');
                 if (json.error) {
-                    console.error(json.error)
+                    alert(json.error);
                 } else {
-                    console.log('Server received valid data');
                     if (json.data) {
-                        //console.log(json.data);
                         props.setReservations(json.data)
                     }
                 }
-            });
+            })
+            .finally(() => {
+                setDisabled(false);
+            })
         }
     }
